Postpone the windowmanager exit-time ,reduce the time waiting for process terminating

+++ ukui-session/modulemanager.cpp
@@ -362,21 +362,31 @@ void ModuleManager::restartModules(int /*exitCode*/, QProcess::ExitStatus exitSt
 void ModuleManager::logout(bool doExit)
 {
     ModulesMapIterator i(mNameMap);
+    UkuiModule *winman;
     while (i.hasNext()) {
         i.next();
         qDebug() << "Module logout" << i.key();
         UkuiModule *p = i.value();
+        if(p->file.name() == QFileInfo(mWindowManager.name()).fileName()){
+            winman = p;
+            continue;
+        }
         p->terminate();
     }
     i.toFront();
     while (i.hasNext()) {
         i.next();
         UkuiModule *p = i.value();
-        if (p->state() != QProcess::NotRunning && !p->waitForFinished(2000)) {
+        if(p->file.name() == QFileInfo(mWindowManager.name()).fileName()){
+            continue;
+        }
+        if (p->state() != QProcess::NotRunning && !p->waitForFinished(200)) {
             qWarning() << "Module " << qPrintable(i.key()) << " won't termiante .. killing.";
             p->kill();
         }
     }
+    winman->terminate();
+
     if (doExit) {
         //QCoreApplication::exit(0);
         exit(0);
